# wp-plugin/shelterapp/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "shelterapp-backend API",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
  "paths" : {
...
    "/animal-procedures" : {
      "get" : {
        "tags" : [ "Animal Procedure Resource" ],
        "parameters" : [ {
          "name" : "animalId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer",
            "nullable" : true
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AnimalProcedure"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Animal Procedure Resource" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AnimalProcedure"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnimalProcedure"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Animal Procedure Resource" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AnimalProcedure"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnimalProcedure"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Animal Procedure Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AnimalProcedure" : {
        "required" : [ "id", "title", "date", "animalId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "allOf" : [ {
              "$ref" : "#/components/schemas/UUID"
            } ],
            "readOnly" : true
          },
          "title" : {
            "pattern" : "\\S",
            "type" : "string"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "animalId" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }
      },
...
      "LocalDate" : {
        "format" : "date",
        "type" : "string",
        "example" : "2022-03-10"
      },
...
      "UUID" : {
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
        "type" : "string"
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "description" : "Authentication",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}